refactor: Move API-related logic to useAuth.ts
