+++ src/pages/Login.tsx
@@ -2,39 +2,26 @@ import { useForm } from "react-hook-form";
 import Title from "../components/common/Title";
 import InputText from "../components/common/InputText";
 import Button from "../components/common/Button";
-import { Link, useNavigate } from "react-router-dom";
-import { login } from "../api/auth.api";
-import { useAlert } from "../hooks/useAlert";
+import { Link } from "react-router-dom";
 import { SignupStyle } from "./Signup";
-import { useAuthStore } from "../store/authStore";
+import { useAuth } from "@/hooks/useAuth";
 
-export interface SignupProps {
+export interface LoginProps {
     email: string;
     password: string;
 }
 
 function Login() {
-    const navigate = useNavigate();
-    const { showAlert } = useAlert();
-
-    const { storeLogin } = useAuthStore();
+    const { userLogin } = useAuth();
 
     const {
         register,
         handleSubmit,
         formState: { errors }
-    } = useForm<SignupProps>();
+    } = useForm<LoginProps>();
 
-    const onSubmit = (data: SignupProps) => {
-        login(data)
-            .then((res) => {
-                storeLogin(res.accessToken);
-                showAlert("로그인 완료");
-                navigate("/");
-            })
-            .catch((error) => {
-                showAlert("로그인 실패");
-            });
+    const onSubmit = (data: LoginProps) => {
+        userLogin(data);
     };
 
     return (
