+++ src/pages/ResetPassword.tsx
@@ -2,11 +2,8 @@ import { useForm } from "react-hook-form";
 import Title from "../components/common/Title";
 import InputText from "../components/common/InputText";
 import Button from "../components/common/Button";
-import { useNavigate } from "react-router-dom";
-import { useState } from "react";
-import { resetPassword, resetRequest } from "../api/auth.api";
-import { useAlert } from "../hooks/useAlert";
 import { SignupStyle } from "./Signup";
+import { useAuth } from "@/hooks/useAuth";
 
 export interface SignupProps {
     email: string;
@@ -14,9 +11,7 @@ export interface SignupProps {
 }
 
 function ResetPassword() {
-    const navigate = useNavigate();
-    const { showAlert } = useAlert();
-    const [resetRequested, setResetRequested] = useState(false);
+    const { resetRequested, userResetPassword, userResetRequest } = useAuth();
 
     const {
         register,
@@ -25,16 +20,7 @@ function ResetPassword() {
     } = useForm<SignupProps>();
 
     const onSubmit = (data: SignupProps) => {
-        if (resetRequested) {
-            resetPassword(data).then(() => {
-                showAlert("비밀번호 초기화 완료");
-                navigate("/login");
-            });
-        } else {
-            resetRequest(data).then(() => {
-                setResetRequested(true);
-            });
-        }
+        resetRequested ? userResetPassword(data) : userResetRequest(data);
     };
 
     return (
