+++ src/pages/Signup.tsx
@@ -3,9 +3,8 @@ import styled from "styled-components";
 import Title from "../components/common/Title";
 import InputText from "../components/common/InputText";
 import Button from "../components/common/Button";
-import { Link, useNavigate } from "react-router-dom";
-import { signup } from "../api/auth.api";
-import { useAlert } from "../hooks/useAlert";
+import { Link } from "react-router-dom";
+import { useAuth } from "@/hooks/useAuth";
 
 export interface SignupProps {
     email: string;
@@ -13,8 +12,7 @@ export interface SignupProps {
 }
 
 function Signup() {
-    const navigate = useNavigate();
-    const { showAlert } = useAlert();
+    const { userSignup } = useAuth();
 
     const {
         register,
@@ -23,10 +21,7 @@ function Signup() {
     } = useForm<SignupProps>();
 
     const onSubmit = (data: SignupProps) => {
-        signup(data).then(() => {
-            showAlert("회원가입 완료");
-            navigate("/login");
-        });
+        userSignup(data);
     };
 
     return (
